@@ -30,8 +30,9 @@
     "element-resize-event": "latest",
     "lodash.debounce": "latest",
     "lodash.isequal": "latest",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "prop-types": "^15.5.6",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3"
   },
   "devDependencies": {},
   "scripts": {}
